Add rule validation for 'one dot per line'
There are various Counters, extracting their sub expressions and then
delgating them to an extractor. This extractor is then delgating the
extracted expressions to the corresponding extractor (built from a
factory). If the expression is a field access or a method invocation, the
access count is increased.
After all Expression nodes were travesed by this system, the reusult is
the number of 'wrong' expressions in the given expressionStatement (the
expression that has an ';' in the end)